requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
azure-core==1.31.0
azure-identity==1.19.0
azure-storage-blob==12.23.1
black==24.8.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
cryptography==43.0.0
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.112.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
isodate==0.7.2
jiter==0.7.0
Mako==1.3.5
MarkupSafe==2.1.5
motor==3.5.1
msal==1.31.0
msal-extensions==1.2.0
mypy-extensions==1.0.0
openai==1.53.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.10.1
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pymongo==4.8.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-multipart==0.0.6
# pywin32==308
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.33
starlette==0.38.4
tomli==2.0.1
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
